Added Score, Total, everything is working

+++ Coin.py
@@ -12,6 +12,7 @@ from constants import DELAY, CELL_SIZE
 class Coin:
     COIN_STATUS = True
     CELL_SIZE = 5
+    SCORE = 0
 
     #CANVAS_POSITION = [13 * 20, 23 * 20]
     def __init__(self, position):
@@ -19,8 +20,10 @@ class Coin:
         self.resized = ImageTk.PhotoImage(self.picture.resize((Coin.CELL_SIZE, Coin.CELL_SIZE), Image.Resampling.LANCZOS))
         self.coin_position = [position[0] * CELL_SIZE, position[1] * CELL_SIZE]
         self.position = [position[1], position[0]]  # МАТЕМАТИЧНА ПОЗИЦІЯ
-        self.score_counter = 0
+
+        self.COIN_STATUS = True
         self.id = None
+        self.observers = []
 
     def draw(self):
 
@@ -30,54 +33,44 @@ class Coin:
     def delete(self):
         CANVAS.delete(self.id)
 
+    def register_observer(self, observer):
+        if observer not in self.observers:
+            self.observers.append(observer)
+            observer.update_score(Coin.SCORE)
+            print ("registered observer")
+
+    def unregister_observer(self, observer):
+        self.observers.remove(observer)
+
+    def notify_observers(self):
+        for observer in self.observers:
+            observer.update_score(Coin.SCORE)
+            print ("notified observers")
+
+
+
+    def status(self):
+        if not self.COIN_STATUS:
+            self.notify_observers()
+
     def change_status(self, pacman):
-        if pacman.position == self.position:
-            Coin.COIN_STATUS = False
+
+        if pacman.position == self.position and self.COIN_STATUS:
+            self.COIN_STATUS = False  # Змінюємо статус на "неактивна"
+            #self.delete()  # Видаляємо монетку з ігрового поля
+            Coin.SCORE += 10  # Збільшуємо рахунок
+            self.notify_observers()
             self.delete()
-            self.score_counter += 10
+
 
 
 def coin_check(pacman, coin_list):
-    print(f"Pacman {pacman.position}")
+    #print(f"Pacman {pacman.position}")
     for coin in coin_list:
         coin.change_status(pacman)
         # print(f"Coin: {coin.position}")
 
     Window.after(DELAY // 30, lambda: coin_check(pacman, coin_list))
-
-
-
-
-if __name__ == "__main__":
-    m = FieldDrawing(FIELD)
-    m.draw_field()
-    #c = Coin((1, 1) )
-    #c.draw()
-    #Window.mainloop()
-
-
-    ex = []
-    for i in range(0, 28):
-        for e in range(0, 30):
-            c = Coin([i, e])  # СПИСОК
-            if FIELD[e][i] == 0 and not (20 > e > 8 and 21 > i > 6):
-                #c.draw()
-                ex.append(c)
-    for t in ex:
-        t.draw()
-
-    vp = VisualPacMan()
-    coin_check(vp, ex)
-    Window.bind('<Left>', vp.move_left)
-    Window.bind('<Right>', vp.move_right)
-    Window.bind('<Up>', vp.move_up)
-    Window.bind('<Down>', vp.move_down)
-    Window.mainloop()
-
-
-
-
-
